+++ src/server/index.js
@@ -31,7 +31,8 @@ const proxyServer = new RammerheadProxy({
     ssl: config.ssl,
     getServerInfo: config.getServerInfo,
     disableLocalStorageSync: config.disableLocalStorageSync,
-    jsCache: config.jsCache
+    jsCache: config.jsCache,
+    disableHttp2: config.disableHttp2
 });
 
 if (config.publicDir) addStaticDirToProxy(proxyServer, config.publicDir);