+++ src/util/fixWebsocket.js
@@ -13,9 +13,18 @@ const respondOnWebSocket = hammerheadWS.respondOnWebSocket;
 hammerheadWS.respondOnWebSocket = function (ctx) {
     ctx.res.on('error', (err) => {
         if (err.code !== 'EPIPE') {
-            console.error('Unknown crash-inducing error:', err);
+            // console.error('Unknown crash-inducing error:', err);
         }
         // cleanup end will automatically be handled by the 'end' listener
     });
+
+    // crashes happen when client wants websocket but destination server says no.
+    // reproduced by setting disableHttp2 to true and going to web.whatsapp.com
+    if (!ctx.destRes.upgrade) {
+        ctx.res.end(`HTTP/${ctx.destRes.httpVersion} ${ctx.destRes.statusCode} ${ctx.destRes.statusMessage}\r\r\n`);
+        ctx.destRes.destroy();
+        return;
+    }
+
     respondOnWebSocket(ctx);
 };